provisioning/cluster: add cluster rolling update with concurrent background polling
