+++ internal/provisioning/cluster/cluster_service_reboot_control_loop_test.go
@@ -3,81 +3,20 @@ package cluster_test
 import (
 	"bytes"
 	"context"
-	"crypto/tls"
 	"errors"
-	"io"
-	"log/slog"
-	"os"
 	"slices"
 	"testing"
 	"time"
 
 	incusosapi "github.com/lxc/incus-os/incus-osd/api"
-	incustls "github.com/lxc/incus/v7/shared/tls"
 	"github.com/stretchr/testify/require"
 
 	"github.com/FuturFusion/operations-center/internal/domain"
-	"github.com/FuturFusion/operations-center/internal/lifecycle"
 	"github.com/FuturFusion/operations-center/internal/provisioning"
 	adapterMock "github.com/FuturFusion/operations-center/internal/provisioning/adapter/mock"
-	provisioningCluster "github.com/FuturFusion/operations-center/internal/provisioning/cluster"
-	serviceMock "github.com/FuturFusion/operations-center/internal/provisioning/mock"
-	"github.com/FuturFusion/operations-center/internal/provisioning/repo/sqlite"
-	"github.com/FuturFusion/operations-center/internal/provisioning/repo/sqlite/entities"
-	provisioningServer "github.com/FuturFusion/operations-center/internal/provisioning/server"
-	"github.com/FuturFusion/operations-center/internal/sql/dbschema"
-	dbdriver "github.com/FuturFusion/operations-center/internal/sql/sqlite"
-	"github.com/FuturFusion/operations-center/internal/sql/transaction"
-	"github.com/FuturFusion/operations-center/internal/util/logger"
 	"github.com/FuturFusion/operations-center/internal/util/ptr"
-	"github.com/FuturFusion/operations-center/internal/util/testing/uuidgen"
 	"github.com/FuturFusion/operations-center/shared/api"
 )
-
-// rebootOnlyServer returns a server, which is fully up to date and does therefore
-// not ask for a reboot on its own. An on demand rolling reboot has to reboot it
-// nevertheless.
-func rebootOnlyServer(t *testing.T, name string) provisioning.Server {
-	t.Helper()
-
-	certPEM, _, err := incustls.GenerateMemCert(false, false)
-	require.NoError(t, err)
-
-	fingerprint, err := incustls.CertFingerprintStr(string(certPEM))
-	require.NoError(t, err)
-
-	return provisioning.Server{
-		Name:          name,
-		Cluster:       new("clusterA"),
-		Type:          api.ServerTypeIncus,
-		ConnectionURL: "https://" + name + "/",
-		Certificate:   string(certPEM),
-		Fingerprint:   fingerprint,
-		HardwareData:  api.HardwareData{},
-		VersionData: api.ServerVersionData{
-			OS: api.OSVersionData{
-				Name:        "incusos",
-				Version:     "1",
-				VersionNext: "1",
-				NeedsReboot: false,
-			},
-			Applications: []api.ApplicationVersionData{
-				{
-					Name:          "incus",
-					Version:       "1",
-					InMaintenance: api.NotInMaintenance,
-				},
-			},
-			NeedsUpdate:   new(false),
-			NeedsReboot:   new(false),
-			InMaintenance: new(api.NotInMaintenance),
-			UpdateChannel: "stable",
-		},
-		Status:       api.ServerStatusReady,
-		StatusDetail: api.ServerStatusDetailNone,
-		Channel:      "stable",
-	}
-}
 
 // rebootOnlyServerClient returns a client mock, which drives the given world
 // through an on demand rolling reboot, i.e. without ever applying an update.
@@ -174,126 +113,18 @@ func rebootOnlyWorld(servers ...provisioning.Server) *serverWorld {
 func setupRebootOnlyCluster(t *testing.T, ctx context.Context, listenerName string, servers ...provisioning.Server) (provisioning.ClusterService, *serverWorld, *bytes.Buffer) {
 	t.Helper()
 
-	certPEM, _, err := incustls.GenerateMemCert(false, false)
-	require.NoError(t, err)
-
-	fingerprint, err := incustls.CertFingerprintStr(string(certPEM))
-	require.NoError(t, err)
-
-	serverNames := make([]string, 0, len(servers))
-	for _, server := range servers {
-		serverNames = append(serverNames, server.Name)
-	}
-
-	clusterA := provisioning.Cluster{
-		Name:          "clusterA",
-		ConnectionURL: "https://cluster-one/",
-		Certificate:   new(string(certPEM)),
-		Fingerprint:   fingerprint,
-		Status:        api.ClusterStatusReady,
-		ServerNames:   serverNames,
-		Channel:       "stable",
-		Config: api.ClusterConfig{
-			RollingRestart: api.ClusterConfigRollingRestart{
-				PostRestoreDelay: (2 * controlLoopInterval).String(),
-			},
-		},
-	}
-
 	world := rebootOnlyWorld(servers...)
-
-	logBuf := &bytes.Buffer{}
-	var logSink io.Writer = logBuf
-	if testing.Verbose() {
-		logSink = io.MultiWriter(os.Stdout, logBuf)
-	}
-
-	err = logger.InitLogger(logSink, "", false, true, true)
-	require.NoError(t, err)
-
-	tmpDir := t.TempDir()
-	db, err := dbdriver.Open(tmpDir)
-	require.NoError(t, err)
-
-	t.Cleanup(func() {
-		err := db.Close()
-		require.NoError(t, err)
-	})
-
-	_, err = dbschema.Ensure(ctx, db, tmpDir)
-	require.NoError(t, err)
-
-	tx := transaction.Enable(db)
-	entities.PreparedStmts, err = entities.PrepareStmts(tx, false)
-	require.NoError(t, err)
-
-	clusterDB := sqlite.NewCluster(tx)
-	serverDB := sqlite.NewServer(tx)
-
-	_, err = clusterDB.Create(ctx, clusterA)
-	require.NoError(t, err)
-
-	for _, server := range servers {
-		_, err = serverDB.Create(ctx, server)
-		require.NoError(t, err)
-	}
-
-	channelSvc := &serviceMock.ChannelServiceMock{
-		GetByNameFunc: func(ctx context.Context, name string) (*provisioning.Channel, error) {
-			return &provisioning.Channel{}, nil
-		},
-	}
 
 	// The most recent available version matches the installed one, nothing is
 	// pending for any of the servers.
-	updateSvc := &serviceMock.UpdateServiceMock{
-		GetAllWithFilterFunc: func(ctx context.Context, filter provisioning.UpdateFilter) (provisioning.Updates, error) {
-			return provisioning.Updates{
-				{
-					ID:      1,
-					UUID:    uuidgen.FromPattern(t, "1"),
-					Version: "1",
-					Files: provisioning.UpdateFiles{
-						{Filename: "x86_64/IncusOS_20260610.img.gz"},
-						{Filename: "x86_64/incus.raw.gz"},
-					},
-				},
-			}, nil
-		},
-	}
-
-	serverSvc := provisioningServer.New(
-		serverDB, rebootOnlyServerClient(world), nil, nil, nil, channelSvc, updateSvc, tls.Certificate{},
-		provisioningServer.WithRebootStatusUpdateGracePeriod(0),
-	)
-
-	clusterSvc := provisioningCluster.New(
-		clusterDB, nil, nil, serverSvc, nil, nil, nil, nil,
-		provisioningCluster.WithPendingUpdateRecheckInterval(controlLoopInterval),
-		provisioningCluster.WithWarningEmitter(provisioning.LogWarningService{}),
-	)
-
-	serverSvc.SetClusterService(clusterSvc)
-
-	// Trigger ClusterUpdateControlLoop also from server lifecycle events.
-	lifecycle.ServerLifecycleSignal.AddListenerWithErr(func(ctx context.Context, slm lifecycle.ServerLifecycleMessage) error {
-		err := clusterSvc.ClusterUpdateControlLoop(ctx, slm.Cluster)
-		if err != nil {
-			slog.ErrorContext(ctx, "Failed to handle server lifecycle event", logger.Err(err), slog.String("server", slm.Server), slog.String("cluster", ptr.From(slm.Cluster)), slog.String("update_state", slm.ServerUpdateState.String()))
-		}
-
-		return err
-	}, listenerName)
-	t.Cleanup(func() {
-		lifecycle.ServerLifecycleSignal.RemoveListener(listenerName)
-	})
+	clusterSvc, _, logBuf := setupControlLoopCluster(t, ctx, listenerName, rebootOnlyServerClient(world), "1", servers...)
 
 	return clusterSvc, world, logBuf
 }
 
 // driveRebootToCompletion runs the control loop until the rolling reboot has
 // finished and returns the progress descriptions, a user would have observed.
-func driveRebootToCompletion(t *testing.T, ctx context.Context, clusterSvc provisioning.ClusterService, world *serverWorld, iterations int) []string {
+func driveRebootToCompletion(t *testing.T, ctx context.Context, clusterSvc provisioning.ClusterService, world *serverWorld, iterations int, beforeIteration ...func(ctx context.Context)) []string {
 	t.Helper()
 
 	var observed []string
@@ -317,6 +148,10 @@ func driveRebootToCompletion(t *testing.T, ctx context.Context, clusterSvc provi
 		// the state, the action leads to, is reported at least once.
 		pending := world.pendingCount()
 
+		for _, hook := range beforeIteration {
+			hook(ctx)
+		}
+
 		err = clusterSvc.ClusterUpdateControlLoop(ctx, nil)
 		if !domain.IsRetryableError(err) {
 			require.NoError(t, err)
@@ -338,7 +173,7 @@ func TestClusterService_ClusterRollingRebootControlLoopSingleNodeCluster(t *test
 	ctx, cancel := context.WithTimeout(t.Context(), asyncActionsDelay*50)
 	defer cancel()
 
-	clusterSvc, world, logBuf := setupRebootOnlyCluster(t, ctx, "ClusterRollingRebootCycleSingleNodeCluster", rebootOnlyServer(t, "one"))
+	clusterSvc, world, logBuf := setupRebootOnlyCluster(t, ctx, "ClusterRollingRebootCycleSingleNodeCluster", clusterMemberServer(t, "one"))
 
 	err := clusterSvc.LaunchClusterReboot(ctx, "clusterA")
 	require.NoError(t, err)
@@ -369,9 +204,9 @@ func TestClusterService_ClusterRollingRebootControlLoopMultiNodeCluster(t *testi
 
 	clusterSvc, world, logBuf := setupRebootOnlyCluster(
 		t, ctx, "ClusterRollingRebootCycleMultiNodeCluster",
-		rebootOnlyServer(t, "serverA"),
-		rebootOnlyServer(t, "serverB"),
-		rebootOnlyServer(t, "serverC"),
+		clusterMemberServer(t, "serverA"),
+		clusterMemberServer(t, "serverB"),
+		clusterMemberServer(t, "serverC"),
 	)
 
 	err := clusterSvc.LaunchClusterReboot(ctx, "clusterA")
@@ -414,7 +249,7 @@ func TestClusterService_ClusterRollingRebootControlLoopMarksServerRebooted(t *te
 	ctx, cancel := context.WithTimeout(t.Context(), asyncActionsDelay*50)
 	defer cancel()
 
-	clusterSvc, world, _ := setupRebootOnlyCluster(t, ctx, "ClusterRollingRebootCycleMarksServerRebooted", rebootOnlyServer(t, "one"))
+	clusterSvc, world, _ := setupRebootOnlyCluster(t, ctx, "ClusterRollingRebootCycleMarksServerRebooted", clusterMemberServer(t, "one"))
 
 	err := clusterSvc.LaunchClusterReboot(ctx, "clusterA")
 	require.NoError(t, err)
@@ -473,7 +308,7 @@ func TestClusterService_LaunchClusterRebootRejectsUnsuitableServers(t *testing.T
 			ctx, cancel := context.WithTimeout(t.Context(), asyncActionsDelay*50)
 			defer cancel()
 
-			server := rebootOnlyServer(t, "one")
+			server := clusterMemberServer(t, "one")
 			tc.mutate(&server)
 
 			clusterSvc, _, _ := setupRebootOnlyCluster(t, ctx, "ClusterRollingRebootReject"+tc.name, server)
@@ -495,7 +330,7 @@ func TestClusterService_LaunchClusterRebootRejectsSecondRun(t *testing.T) {
 	ctx, cancel := context.WithTimeout(t.Context(), asyncActionsDelay*50)
 	defer cancel()
 
-	clusterSvc, _, _ := setupRebootOnlyCluster(t, ctx, "ClusterRollingRebootRejectsSecondRun", rebootOnlyServer(t, "one"))
+	clusterSvc, _, _ := setupRebootOnlyCluster(t, ctx, "ClusterRollingRebootRejectsSecondRun", clusterMemberServer(t, "one"))
 
 	err := clusterSvc.LaunchClusterReboot(ctx, "clusterA")
 	require.NoError(t, err)
